feat: [M1957] Fix typing, comments, file location
