+++ src/components/generic/GatedFeature/index.js
@@ -1,0 +1,3 @@
+import GatedFeature from './GatedFeature'
+
+export default GatedFeature
